Flatten styles in test snapshots
