+++ .travis.yml
@@ -6,4 +6,3 @@ before_script:
   - sh -e /etc/init.d/xvfb start
 script:
   - yarn lint
-  - yarn test:ci
